refactored mem-store into /store
