+++ mender_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/mendersoftware/mender/client"
 	cltest "github.com/mendersoftware/mender/client/test"
 	"github.com/mendersoftware/mender/store"
-	"github.com/mendersoftware/mender/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
@@ -96,7 +95,7 @@ func newTestMender(runner *testOSCalls, config menderConfig, pieces testMenderPi
 	// fill out missing pieces
 
 	if pieces.store == nil {
-		pieces.store = utils.NewMemStore()
+		pieces.store = store.NewMemStore()
 	}
 
 	if pieces.device == nil {
@@ -130,7 +129,7 @@ func newDefaultTestMender() *mender {
 
 func Test_ForceBootstrap(t *testing.T) {
 	// generate valid keys
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			DeviceKey: "temp.key",
@@ -185,7 +184,7 @@ func Test_Bootstrap(t *testing.T) {
 func Test_BootstrappedHaveKeys(t *testing.T) {
 
 	// generate valid keys
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	k := store.NewKeystore(ms, "temp.key")
 	assert.NotNil(t, k)
 	assert.NoError(t, k.Generate())
@@ -212,7 +211,7 @@ func Test_BootstrappedHaveKeys(t *testing.T) {
 
 func Test_BootstrapError(t *testing.T) {
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 
 	ms.Disable(true)
 
@@ -500,7 +499,7 @@ func TestMenderReportStatus(t *testing.T) {
 	srv := cltest.NewClientTestServer()
 	defer srv.Close()
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			ServerURL: srv.URL,
@@ -562,7 +561,7 @@ func TestMenderLogUpload(t *testing.T) {
 	srv := cltest.NewClientTestServer()
 	defer srv.Close()
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			ServerURL: srv.URL,
@@ -641,7 +640,7 @@ func TestAuthToken(t *testing.T) {
 	ts := cltest.NewClientTestServer()
 	defer ts.Close()
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			ServerURL: ts.URL,
@@ -682,7 +681,7 @@ func TestMenderInventoryRefresh(t *testing.T) {
 	srv := cltest.NewClientTestServer()
 	defer srv.Close()
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			ServerURL: srv.URL,
@@ -947,7 +946,7 @@ func TestMenderFetchUpdate(t *testing.T) {
 
 	srv.Update.Has = true
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	mender := newTestMender(nil,
 		menderConfig{
 			ServerURL: srv.URL,
