+++ state_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/mendersoftware/log"
 	"github.com/mendersoftware/mender/client"
-	"github.com/mendersoftware/mender/utils"
+	"github.com/mendersoftware/mender/store"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -240,7 +240,7 @@ func TestStateUpdateError(t *testing.T) {
 	assert.NotNil(t, errstate)
 	assert.Equal(t, fooerr, errstate.cause)
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -259,7 +259,7 @@ func TestStateUpdateReportStatus(t *testing.T) {
 		ID: "foobar",
 	}
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -418,7 +418,7 @@ func TestStateAuthorized(t *testing.T) {
 	var s State
 	var c bool
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -617,7 +617,7 @@ func TestStateUpdateCommit(t *testing.T) {
 	var s State
 	var c bool
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -745,7 +745,7 @@ func TestStateUpdateFetch(t *testing.T) {
 	}
 	cs := NewUpdateFetchState(update)
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -869,7 +869,7 @@ func TestStateUpdateFetchRetry(t *testing.T) {
 		ID: "foobar",
 	}
 	cs := NewUpdateFetchState(update)
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -929,7 +929,7 @@ func TestStateUpdateInstall(t *testing.T) {
 	}
 	uis := NewUpdateInstallState(stream, int64(len(data)), update)
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -977,7 +977,7 @@ func TestStateUpdateInstallRetry(t *testing.T) {
 	data := "test"
 	stream := ioutil.NopCloser(bytes.NewBufferString(data))
 	uis := NewUpdateInstallState(stream, int64(len(data)), update)
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -1042,7 +1042,7 @@ func TestStateReboot(t *testing.T) {
 	defer os.RemoveAll(tempDir)
 	DeploymentLogger = NewDeploymentLogManager(tempDir)
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := StateContext{
 		store: ms,
 	}
@@ -1114,7 +1114,7 @@ func TestStateFinal(t *testing.T) {
 }
 
 func TestStateData(t *testing.T) {
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	sd := StateData{
 		Version: stateDataVersion,
 		Name:    MenderStateInit,
@@ -1152,7 +1152,7 @@ func TestStateReportError(t *testing.T) {
 		ID: "foobar",
 	}
 
-	ms := utils.NewMemStore()
+	ms := store.NewMemStore()
 	ctx := &StateContext{
 		store: ms,
 	}
