+++ store/storeMock.go
@@ -14,31 +14,10 @@
 package store
 
 import (
-	"bytes"
 	"io"
 
-	"github.com/mendersoftware/mender/utils"
 	"github.com/stretchr/testify/mock"
 )
-
-type MemStoreWriter struct {
-	bytes.Buffer
-	ms   *utils.MemStore
-	name string
-}
-
-func (msw *MemStoreWriter) Commit() error {
-	msw.ms.Commit(msw.name, msw.Bytes())
-	return nil
-}
-
-func (msw *MemStoreWriter) Close() error {
-	return nil
-}
-
-type MemStoreData struct {
-	data []byte
-}
 
 type MockStore struct {
 	mock.Mock
@@ -64,13 +43,13 @@ func (ms *MockStore) Close() error {
 	return ret.Error(0)
 }
 
-func (ms *MockStore) OpenWrite(name string) (utils.WriteCloserCommitter, error) {
+func (ms *MockStore) OpenWrite(name string) (WriteCloserCommitter, error) {
 	ret := ms.Called(name)
 	wcc := ret.Get(0)
 	if wcc == nil {
 		return nil, ret.Error(1)
 	}
-	return ret.Get(0).(utils.WriteCloserCommitter), ret.Error(1)
+	return ret.Get(0).(WriteCloserCommitter), ret.Error(1)
 }
 
 func (ms *MockStore) OpenRead(name string) (io.ReadCloser, error) {
